+++ src/mocks/main.jsx
@@ -1,0 +1,13 @@
+const titles = [
+  { title: 'LEAD_1', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_2', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_3', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_4', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_5', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_6', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_7', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_8', content: '김세론a, 김세론, 김세론c' },
+  { title: 'LEAD_9', content: '김세론a, 김세론, 김세론c' },
+];
+
+export default titles;
